requirements.txt
fastapi
uvicorn[standard]
gunicorn
websockets
python-dotenv
